core/requirements.txt
asgiref==3.8.1
Django==5.1.6
djangorestframework==3.15.2
mysqlclient==2.2.7
numpy==2.2.3
pyarrow==19.0.1
pylance==0.23.0
sqlparse==0.5.3
